Remove String#first and String#last
Active Support >= 6.1 will raise an exception when a negative `limit` is
passed to String#first or String#last (see rails/rails#29105).  Although
it is debatable whether it would be better to return an empty string as
these implementations do, the behavior present in Active Support <= 6.0
will no longer need to be corrected.

Additionally, the performance optimizations have been contributed to
Active Support in rails/rails#36185.